back-end/BoundBoxApi/requirement_pip.txt
# pip install -r requirement_pip.txt
django == 2.2.5
djangorestframework == 3.10.3
django-filter == 2.2.0
django-cors-headers == 3.1.0
psycopg2-binary == 2.8.3
numpy == 1.17.2
torch == 1.2.0
torchvision == 0.4.0